refactor: make my interests and notification preference e2es viewport aware
