+++ projects/storefrontapp-e2e-cypress/cypress/helpers/notification.ts
@@ -15,11 +15,6 @@ export function navigateToNotificationPreferencePage() {
   });
 }
 
-export function verifyNotificationPrefAsAnonymous() {
-  cy.visit('/my-account/notification-preference');
-  cy.location('pathname').should('contain', '/login');
-}
-
 export function enableNotificationChannel() {
   navigateToNotificationPreferencePage();
   cy.get('[type="checkbox"]').first().check();
@@ -28,13 +23,6 @@ export function enableNotificationChannel() {
 export function disableNotificationChannel() {
   navigateToNotificationPreferencePage();
   cy.get('[type="checkbox"]').first().uncheck();
-}
-
-export function verifyNotificationChannel() {
-  enableNotificationChannel();
-  verifyChannelEnabled();
-  disableNotificationChannel();
-  verifyChannelDisabled();
 }
 
 export function updateEmail(): String {
@@ -49,24 +37,6 @@ export function updateEmail(): String {
   cy.get('cx-update-email-form button[type="submit"]').click();
   login(newUid, password);
   return newUid;
-}
-
-export function verifyChannelValueUpdating() {
-  verifyEmailChannel(standardUser.registrationData.email);
-  const newEmail = updateEmail();
-  verifyEmailChannel(newEmail);
-}
-
-export function verifyChannelDisabled() {
-  cy.visit('/');
-  navigateToNotificationPreferencePage();
-  cy.get('[type="checkbox"]').first().should('not.be.checked');
-}
-
-export function verifyChannelEnabled() {
-  cy.visit('/');
-  navigateToNotificationPreferencePage();
-  cy.get('[type="checkbox"]').first().should('be.checked');
 }
 
 export function verifyEmailChannel(email: String) {
@@ -148,11 +118,6 @@ export function verifyNavigateToNotificationPreferenceInDialog() {
 export function verifyNavigateToMyInterestsInDialog() {
   cy.get('.link-interests').click();
   verifyCustomerInterest(normalProductCode);
-}
-//Customer interest
-export function verifyMyInterestsAsAnonymous() {
-  cy.visit('/my-account/my-interests');
-  cy.location('pathname').should('contain', '/login');
 }
 
 export function verifySubscriptionAndCustomerInterest(productCode: string) {
