+++ projects/storefrontapp-e2e-cypress/cypress/integration/regression/notification/my-interests.e2e-spec.ts
@@ -1,41 +1,47 @@
 import * as notification from '../../../helpers/notification';
 import * as orderDetail from '../../../helpers/consignment-tracking';
+import { viewportContext } from '../../../helpers/viewport-context';
 
-describe('my interests - guest', () => {
-  before(() => {
-    cy.window().then((win) => win.sessionStorage.clear());
-  });
+describe('My interests', () => {
+  viewportContext(['mobile', 'desktop'], () => {
+    describe('Guest', () => {
+      before(() => {
+        cy.window().then((win) => win.sessionStorage.clear());
+      });
 
-  it('should redirect to login page for anonymous user', () => {
-    notification.verifyMyInterestsAsAnonymous();
-  });
-});
+      it('should redirect to login page for anonymous user', () => {
+        cy.visit('/my-account/my-interests');
+        cy.location('pathname').should('contain', '/login');
+      });
+    });
 
-describe('my interests - customer', () => {
-  beforeEach(() => {
-    cy.window().then((win) => win.sessionStorage.clear());
-    cy.requireLoggedIn();
-    cy.visit('/');
-    notification.enableNotificationChannel();
-  });
+    describe('Customer', () => {
+      beforeEach(() => {
+        cy.window().then((win) => win.sessionStorage.clear());
+        cy.requireLoggedIn();
+        cy.visit('/');
+        notification.enableNotificationChannel();
+      });
 
-  it('should show/remove interest', () => {
-    notification.verifyCustomerInterests();
-  });
+      it('should show/remove interest', () => {
+        notification.verifyCustomerInterests();
+      });
 
-  it('should remove the subscrption in PDP', () => {
-    notification.verifyRemovingCustomerInterestInPDP();
-  });
-});
+      it('should remove the subscrption in PDP', () => {
+        notification.verifyRemovingCustomerInterestInPDP();
+      });
+    });
 
-describe('my interests - customer with interests', () => {
-  before(() => {
-    cy.window().then((win) => win.sessionStorage.clear());
-    cy.requireLoggedIn();
-    orderDetail.loginUsingUserWithOrder();
-  });
+    describe('Customer with interests', () => {
+      before(() => {
+        cy.window().then((win) => win.sessionStorage.clear());
+        cy.requireLoggedIn();
+        orderDetail.loginUsingUserWithOrder();
+      });
 
-  it('should page and sort', () => {
-    notification.verifyPagingAndSorting();
+      it('should page and sort', () => {
+        notification.verifyPagingAndSorting();
+      });
+    });
   });
 });
